+++ src/com/herocraftonline/dev/heroes/util/ConfigManager.java
@@ -125,7 +125,7 @@ public class ConfigManager {
         if (killing != null) {
             for (String item : killing) {
                 try {
-                    int exp = config.getInt(root + "." + item, 0);
+                    double exp = config.getDouble(root + "." + item, 0);
                     if (item.equals("player")) {
                         properties.playerKillingExp = exp;
                     } else {
@@ -142,7 +142,7 @@ public class ConfigManager {
         List<String> mining = config.getKeys(root);
         if (mining != null) {
             for (String item : mining) {
-                int exp = config.getInt(root + "." + item, 0);
+                double exp = config.getDouble(root + "." + item, 0);
                 Material type = Material.matchMaterial(item);
 
                 if (type != null) {
@@ -157,7 +157,7 @@ public class ConfigManager {
         List<String> logging = config.getKeys(root);
         if (logging != null) {
             for (String item : logging) {
-                int exp = config.getInt(root + "." + item, 0);
+                double exp = config.getDouble(root + "." + item, 0);
                 Material type = Material.matchMaterial(item);
 
                 if (type != null) {
@@ -172,7 +172,7 @@ public class ConfigManager {
         List<String> crafting = config.getKeys(root);
         if (crafting != null) {
             for (String item : crafting) {
-                int exp = config.getInt(root + "." + item, 0);
+                double exp = config.getDouble(root + "." + item, 0);
                 Material type = Material.matchMaterial(item);
 
                 if (type != null) {
